add linting&formatting

+++ contracts/interface/ICurve1.sol
@@ -3,15 +3,17 @@ pragma solidity >=0.7.0;
 
 interface ICurve1 {
     function underlying_coins(uint256 index) external view returns (address);
+
     function exchange_underlying(
         uint256 tokenIndexFrom,
         uint256 tokenIndexTo,
         uint256 dx,
         uint256 minDy
     ) external;
+
     function get_dy_underlying(
-        uint tokenIndexFrom,
-        uint tokenIndexTo,
+        uint256 tokenIndexFrom,
+        uint256 tokenIndexTo,
         uint256 dx
     ) external view returns (uint256);
 }
